+++ palace/models/timeoperator.hpp
@@ -24,41 +24,38 @@ class TimeOperator
 {
 private:
   // Solution vector storage.
-  Vector E, dE, En, B;
+  Vector E, B, sol;
 
-  // Time integrator for the curl-curl E-field formulation.
-  std::unique_ptr<mfem::SecondOrderODESolver> ode;
+  // Time integrator for the first order ODE system.
+  std::unique_ptr<mfem::ODESolver> ode;
 
-  // Time-dependent operator for the E-field.
-  std::unique_ptr<mfem::SecondOrderTimeDependentOperator> op;
+  // Time-dependent operator for the Edot-E ODE system.
+  std::unique_ptr<mfem::TimeDependentOperator> op;
 
-  // Discrete curl for B-field time integration (not owned).
-  const Operator *Curl;
+  // Adaptive time-stepping parameters.
+  int order;
+  double rel_tol, abs_tol;
+  bool use_mfem_integrator = false;
 
 public:
   TimeOperator(const IoData &iodata, SpaceOperator &space_op,
-               std::function<double(double)> &dJ_coef);
+               std::function<double(double)> dJ_coef);
 
   // Access solution vectors for E- and B-fields.
   const Vector &GetE() const { return E; }
-  const Vector &GetEdot() const { return dE; }
   const Vector &GetB() const { return B; }
 
   // Return the linear solver associated with the implicit or explicit time integrator.
   const KspSolver &GetLinearSolver() const;
-
-  // Return if the time integration scheme explicit or implicit.
-  bool isExplicit() const { return op->isExplicit(); }
-
-  // Estimate the maximum stable time step based on the maximum eigenvalue of the
-  // undamped system matrix M⁻¹ K.
-  double GetMaxTimeStep() const;
 
   // Initialize time integrators and set 0 initial conditions.
   void Init();
 
   // Perform time step from t -> t + dt.
   void Step(double &t, double &dt);
+
+  // Print ODE integrator statistics.
+  void PrintStats();
 };
 
 }  // namespace palace